Moved the tests that were accidentally left in the SEM package to the test package.
