+++ src/test/java/edu/cmu/tetrad/test/TestDagScorer.java
@@ -19,65 +19,62 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA //
 ///////////////////////////////////////////////////////////////////////////////
 
-package edu.cmu.tetrad.sem;
+package edu.cmu.tetrad.test;
 
-import edu.cmu.tetrad.graph.EdgeListGraph;
+import edu.cmu.tetrad.data.DataSet;
+import edu.cmu.tetrad.graph.Dag;
 import edu.cmu.tetrad.graph.Graph;
-import edu.cmu.tetrad.graph.GraphNode;
-import edu.cmu.tetrad.graph.Node;
+import edu.cmu.tetrad.graph.GraphUtils;
+import edu.cmu.tetrad.sem.*;
 import junit.framework.Test;
 import junit.framework.TestCase;
 import junit.framework.TestSuite;
 
-import java.util.List;
 
 /**
+ * Tests the MeasurementSimulator class using diagnostics devised by Richard
+ * Scheines. The diagnostics are described in the Javadocs, below.
+ *
  * @author Joseph Ramsey
  */
-public class TestSemUpdater extends TestCase {
+public class TestDagScorer extends TestCase {
 
     /**
      * Standard constructor for JUnit test cases.
      */
-    public TestSemUpdater(String name) {
+    public TestDagScorer(String name) {
         super(name);
     }
 
-    public void testUpdate() {
-        Graph graph = constructGraph1();
+    public void test1() {
+        Graph dag = new Dag(GraphUtils.randomGraph(10, 0, 10, 30, 15, 15, false));
 
-        SemPm semPm = new SemPm(graph);
-        SemIm semIm = new SemIm(semPm);
+        SemPm pm = new SemPm(dag);
+        SemIm im = new SemIm(pm);
+        DataSet data = im.simulateData(1000, false);
 
-        System.out.println("Original graph: " + semIm.getSemPm().getGraph());
-        System.out.println("Original semIm: " + semIm);
+        GraphUtils.replaceNodes(dag, data.getVariables());
 
-        List nodes = semIm.getVariableNodes();
+        SemEstimator est = new SemEstimator(data, pm);
+        SemIm estSem = est.estimate();
+        System.out.println("FML = " + estSem.getScore());
 
-        SemUpdater semUpdater = new SemUpdater(semIm);
+        dag = GraphUtils.replaceNodes(dag, data.getVariables());
 
-        SemEvidence evidence = new SemEvidence(semIm);
-        evidence.getProposition().setValue((Node) nodes.get(4), 10.0);
-//        evidence.getProposition().setValue((Node) nodes.get(2), 1.5);
+        System.out.println(estSem.getEdgeCoef());
+        System.out.println(estSem.getErrCovar());
 
-        semUpdater.setEvidence(evidence);
+        Scorer scorer = new DagScorer(data);
+        double fml = scorer.score(dag);
+        System.out.println("FML (scorer) = " + fml);
 
+        System.out.println("BIC = " + scorer.getBicScore());
+        System.out.println("DOF = " + scorer.getDof());
+        System.out.println("# free params = " + scorer.getNumFreeParams());
+        System.out.println("est sem = " + scorer.getEstSem());
 
-        evidence.setManipulated(1, true);
-
-        SemIm manipulatedSemIm = semUpdater.getManipulatedSemIm();
-
-        System.out.println("Manipulated Sem IM: " + manipulatedSemIm);
-
-        System.out.println(
-                "Manipulated graph: " + manipulatedSemIm.getSemPm().getGraph());
-        System.out.println("Manipulated semIm: " + manipulatedSemIm);
-
-        SemIm updatedSemIm = semUpdater.getUpdatedSemIm();
-
-        System.out.println(
-                "Updated graph: " + updatedSemIm.getSemPm().getGraph());
-        System.out.println("Updated SEM = " + updatedSemIm);
+        Graph newDag = new Dag(GraphUtils.randomDag(data.getVariables(), 0, 10, 30, 15, 15, false));
+        System.out.println("new FML " + scorer.score(newDag));
     }
 
     /**
@@ -88,35 +85,8 @@ public class TestSemUpdater extends TestCase {
 
         // Edit the name of the class in the parens to match the name
         // of this class.
-        return new TestSuite(TestSemUpdater.class);
-    }
-
-    private Graph constructGraph1() {
-        Graph graph = new EdgeListGraph();
-
-        Node x1 = new GraphNode("X1");
-        Node x2 = new GraphNode("X2");
-        Node x3 = new GraphNode("X3");
-        Node x4 = new GraphNode("X4");
-        Node x5 = new GraphNode("X5");
-
-        graph.addNode(x1);
-        graph.addNode(x2);
-        graph.addNode(x3);
-        graph.addNode(x4);
-        graph.addNode(x5);
-
-        graph.addDirectedEdge(x1, x2);
-        graph.addDirectedEdge(x2, x3);
-        graph.addDirectedEdge(x3, x4);
-        graph.addDirectedEdge(x1, x4);
-        graph.addDirectedEdge(x4, x5);
-
-        return graph;
+        return new TestSuite(TestDagScorer.class);
     }
 }
-
-
-
