+++ src/test/java/edu/cmu/tetrad/test/TestSemXml.java
@@ -19,11 +19,15 @@
 // Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA //
 ///////////////////////////////////////////////////////////////////////////////
 
-package edu.cmu.tetrad.sem;
+package edu.cmu.tetrad.test;
 
 import edu.cmu.tetrad.graph.Dag;
 import edu.cmu.tetrad.graph.Graph;
 import edu.cmu.tetrad.graph.GraphUtils;
+import edu.cmu.tetrad.sem.SemIm;
+import edu.cmu.tetrad.sem.SemPm;
+import edu.cmu.tetrad.sem.SemXmlParser;
+import edu.cmu.tetrad.sem.SemXmlRenderer;
 import junit.framework.Test;
 import junit.framework.TestCase;
 import junit.framework.TestSuite;
